requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
dataclasses-json==0.6.4
deepgram-sdk==3.2.6
Deprecated==1.2.14
exceptiongroup==1.2.0
fastapi==0.109.0
Flask==3.0.0
flask-sock==0.7.0
frozenlist==1.4.1
gevent==23.9.1
greenlet==3.0.3
gTTS==2.5.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.35.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
verboselogs==1.7
vonage==3.13.0
vonage-jwt==1.1.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
